requirements.txt
requests==2.31.0
beautifulsoup4~=4.9.0
pymongo~=3.11.4
Unidecode~=1.1.1
psycopg2
fake-useragent